single deactivate function by full key
