+++ lambda_functions/v1/tests/code_generator/conftest.py
@@ -5,6 +5,7 @@ import pytest
 from moto import mock_dynamodb2
 
 from lambda_functions.v1.functions.lpa_codes.app.api import code_generator
+from lambda_functions.v1.tests.code_generator.test_data import test_data
 
 
 @pytest.fixture()
@@ -67,11 +68,6 @@ def mock_database():
         table.update(GlobalSecondaryIndexUpdates=code_index)
         table.update(GlobalSecondaryIndexUpdates=active_index)
 
-        # with open("test_data.json") as data_file:
-        #     data = json.load(data_file)
-        #     for row in data:
-        #         table.put_item(Item=row)
-
         data = test_data()
         for row in data:
             table.put_item(Item=row)
@@ -80,71 +76,3 @@ def mock_database():
 
         print("db teardown")
         table.delete()
-
-
-def test_data():
-    return [
-        {
-            "active": True,
-            "actor": "mediumblue",
-            "code": "YsSu4iAztUXm",
-            "expiry_date": "29/03/2021",
-            "generated_date": "29/03/2020",
-            "last_updated_date": "25/08/2020",
-            "lpa": "drive_leading-edge_communities",
-        },
-        {
-            "active": True,
-            "actor": "mediumblue",
-            "code": "aEYVS6i9zSwy",
-            "expiry_date": "26/06/2020",
-            "generated_date": "27/06/2019",
-            "last_updated_date": "03/02/2020",
-            "lpa": "drive_leading-edge_communities",
-        },
-        {
-            "active": True,
-            "actor": "mediumblue",
-            "code": "ZY577rXcRVLY",
-            "expiry_date": "05/04/2021",
-            "generated_date": "05/04/2020",
-            "last_updated_date": "28/02/2021",
-            "lpa": "drive_leading-edge_communities",
-        },
-        {
-            "active": False,
-            "actor": "mediumblue",
-            "code": "hFCarGyJF6G2",
-            "expiry_date": "06/08/2020",
-            "generated_date": "07/08/2019",
-            "last_updated_date": "24/03/2020",
-            "lpa": "drive_leading-edge_communities",
-        },
-        {
-            "active": False,
-            "actor": "mediumblue",
-            "code": "hm8Qtyb763tD",
-            "expiry_date": "05/08/2020",
-            "generated_date": "06/08/2019",
-            "last_updated_date": "01/12/2019",
-            "lpa": "drive_leading-edge_communities",
-        },
-        {
-            "active": False,
-            "actor": "mediumblue",
-            "code": "HiRqUNXRKB3X",
-            "expiry_date": "02/01/2021",
-            "generated_date": "03/01/2020",
-            "last_updated_date": "11/04/2020",
-            "lpa": "drive_leading-edge_communities",
-        },
-        {
-            "active": False,
-            "actor": "mediumblue",
-            "code": "UEW7zSi42bLF",
-            "expiry_date": "09/03/2021",
-            "generated_date": "09/03/2020",
-            "last_updated_date": "17/10/2020",
-            "lpa": "drive_leading-edge_communities",
-        },
-    ]
