--- lambda_functions/v1/tests/code_generator/test_data.json
@@ -1,57 +1,0 @@
-[{
-	"lpa": "drive_leading-edge_communities",
-	"actor": "mediumblue",
-	"code": "YsSu4iAztUXm",
-	"active": true,
-	"generated_date": "29/03/2020",
-	"expiry_date": "29/03/2021",
-	"last_updated_date": "25/08/2020"
-}, {
-	"lpa": "drive_leading-edge_communities",
-	"actor": "mediumblue",
-	"code": "aEYVS6i9zSwy",
-	"active": true,
-	"generated_date": "27/06/2019",
-	"expiry_date": "26/06/2020",
-	"last_updated_date": "03/02/2020"
-}, {
-	"lpa": "drive_leading-edge_communities",
-	"actor": "mediumblue",
-	"code": "ZY577rXcRVLY",
-	"active": true,
-	"generated_date": "05/04/2020",
-	"expiry_date": "05/04/2021",
-	"last_updated_date": "28/02/2021"
-}, {
-	"lpa": "drive_leading-edge_communities",
-	"actor": "mediumblue",
-	"code": "hFCarGyJF6G2",
-	"active": false,
-	"generated_date": "07/08/2019",
-	"expiry_date": "06/08/2020",
-	"last_updated_date": "24/03/2020"
-}, {
-	"lpa": "drive_leading-edge_communities",
-	"actor": "mediumblue",
-	"code": "hm8Qtyb763tD",
-	"active": false,
-	"generated_date": "06/08/2019",
-	"expiry_date": "05/08/2020",
-	"last_updated_date": "01/12/2019"
-}, {
-	"lpa": "drive_leading-edge_communities",
-	"actor": "mediumblue",
-	"code": "HiRqUNXRKB3X",
-	"active": false,
-	"generated_date": "03/01/2020",
-	"expiry_date": "02/01/2021",
-	"last_updated_date": "11/04/2020"
-}, {
-	"lpa": "drive_leading-edge_communities",
-	"actor": "mediumblue",
-	"code": "UEW7zSi42bLF",
-	"active": false,
-	"generated_date": "09/03/2020",
-	"expiry_date": "09/03/2021",
-	"last_updated_date": "17/10/2020"
-}]
